+++ node-website/src/components/BackendInterest.js
@@ -1,8 +1,8 @@
 import React from 'react';
 
 export const BackendInterest = () => {
-  return (
-  <div>
-      
+  return (<div>
+      <p>Backend designs are a challenge I was first to pick up on in web. Not too far from my research passion lies the desire to take what I've learned about decoupled class design and apply it to what I do in web.</p> 
+      <p>I've built lots of apps offline in Java, C#, and even Python, and I hook them into Postgres and MySQL databases via sites built using Spring and .NET framework, with a growing interest in working with Django. I need more experience in this area, so contact me if you're hiring!</p>
   </div>);
 };
