+++ node-website/src/components/FrontendInterest.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+export class FrontendInterest extends Component {
+  render() {
+    return <div></div>;
+  }
+}
+
+export default FrontendInterest;
